refactor: 훅 테스트 파일 위치 변경
