+++ src/hooks/tests/useFilteredProducts.test.ts
@@ -3,8 +3,8 @@ import { Category } from '@appTypes/index';
 import { CATEGORY_OPTIONS } from '@constants/index';
 import { renderHook, waitFor } from '@testing-library/react';
 
-import useFetch from './useFetch';
-import useFilteredProducts from './useFilteredProducts';
+import useFetch from '../useFetch';
+import useFilteredProducts from '../useFilteredProducts';
 
 describe('상품 목록 필터링 테스트', () => {
   describe('카테고리 필터링 테스트', () => {
